+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
-using Restaurants.Application.Restaurants.Commands.CreateRestaurant;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.IRepos;
+using Restraurants.Domain.Exceptions;
+using Restraurants.Domain.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,11 +18,14 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
         private IRestaurantsRepo _restaurantsRepo;
         private ILogger<DeleteRestaurantCommandHandler> _logger;
         private IMapper _mapper;
+        private IRestaurantAuthorizationService _restaurantAuthorizationService;
         public DeleteRestaurantCommandHandler(IRestaurantsRepo restaurantsRepo,
             ILogger<DeleteRestaurantCommandHandler> logger,
+            IRestaurantAuthorizationService restaurantAuthorizationService,
             IMapper mapper)
         {
             _restaurantsRepo = restaurantsRepo;
+            _restaurantAuthorizationService = restaurantAuthorizationService;
             _logger = logger;
             _mapper = mapper;
         }
@@ -30,6 +35,8 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             var restaurant = await _restaurantsRepo.GetByIdAsync(request.Id);
             if (restaurant == null) 
                 return false;
+            if (!_restaurantAuthorizationService.Authorize(restaurant, ResourceOperations.Delete))
+                throw new ForbidException();
             await _restaurantsRepo.Delete(restaurant);
             return true;
         }
